Refactor deleteCompleted function into status.js
